complete memory optimizations, bumped up iterations for k means clustering

+++ src/main/java/com/github/retrooper/bigdata/model/Prediction.java
@@ -1,5 +1,5 @@
 package com.github.retrooper.bigdata.model;
 
 public interface Prediction<T> {
-    double predict(T x);
+    float predict(T x);
 }
